refactor sql eventstore
