+++ eventstore/sql/sql.go
@@ -11,7 +11,7 @@ import (
 	"github.com/hallgren/eventsourcing/eventstore"
 )
 
-// SQL for store events
+// SQL event store handler
 type SQL struct {
 	db         *sql.DB
 	serializer eventsourcing.Serializer
@@ -92,63 +92,6 @@ func (s *SQL) Save(events []eventsourcing.Event) error {
 		events[i].GlobalVersion = eventsourcing.Version(lastInsertedID)
 	}
 	return tx.Commit()
-}
-
-type iterator struct {
-	rows       *sql.Rows
-	serializer eventsourcing.Serializer
-}
-
-func (i *iterator) Close() {
-	i.rows.Close()
-}
-
-func (i *iterator) Next() (eventsourcing.Event, error) {
-	var globalVersion eventsourcing.Version
-	var eventMetadata map[string]interface{}
-	var version eventsourcing.Version
-	var id, reason, typ, timestamp string
-	var data, metadata string
-	if !i.rows.Next() {
-		return eventsourcing.Event{}, eventsourcing.ErrNoMoreEvents
-	}
-	if err := i.rows.Scan(&globalVersion, &id, &version, &reason, &typ, &timestamp, &data, &metadata); err != nil {
-		return eventsourcing.Event{}, err
-	}
-
-	t, err := time.Parse(time.RFC3339, timestamp)
-	if err != nil {
-		return eventsourcing.Event{}, err
-	}
-
-	f, ok := i.serializer.Type(typ, reason)
-	if !ok {
-		// if the typ/reason is not register jump over the event
-		return i.Next()
-	}
-
-	eventData := f()
-	err = i.serializer.Unmarshal([]byte(data), &eventData)
-	if err != nil {
-		return eventsourcing.Event{}, err
-	}
-	if metadata != "" {
-		err = i.serializer.Unmarshal([]byte(metadata), &eventMetadata)
-		if err != nil {
-			return eventsourcing.Event{}, err
-		}
-	}
-
-	event := eventsourcing.Event{
-		AggregateID:   id,
-		Version:       version,
-		GlobalVersion: globalVersion,
-		AggregateType: typ,
-		Timestamp:     t,
-		Data:          eventData,
-		Metadata:      eventMetadata,
-	}
-	return event, nil
 }
 
 // Get the events from database
